+++ shiny/ui/_chat_normalize.py
@@ -2,7 +2,7 @@ import sys
 from abc import ABC, abstractmethod
 from typing import TYPE_CHECKING, Any, Optional, cast
 
-from htmltools import HTML
+from htmltools import HTML, Tagifiable
 
 from ._chat_types import ChatMessage
 
@@ -75,6 +75,22 @@ class DictNormalizer(BaseMessageNormalizer):
 
     def can_normalize_chunk(self, chunk: Any) -> bool:
         return isinstance(chunk, dict)
+
+
+class TagifiableNormalizer(DictNormalizer):
+    def normalize(self, message: Any) -> ChatMessage:
+        x = cast("Tagifiable", message)
+        return super().normalize({"content": x})
+
+    def normalize_chunk(self, chunk: Any) -> ChatMessage:
+        x = cast("Tagifiable", chunk)
+        return super().normalize_chunk({"content": x})
+
+    def can_normalize(self, message: Any) -> bool:
+        return isinstance(message, Tagifiable)
+
+    def can_normalize_chunk(self, chunk: Any) -> bool:
+        return isinstance(chunk, Tagifiable)
 
 
 class LangChainNormalizer(BaseMessageNormalizer):
@@ -259,6 +275,7 @@ class NormalizerRegistry:
             "google": GoogleNormalizer(),
             "langchain": LangChainNormalizer(),
             "ollama": OllamaNormalizer(),
+            "tagify": TagifiableNormalizer(),
             "dict": DictNormalizer(),
             "string": StringNormalizer(),
         }